balance done

+++ internal/repository/migrations/002_create_order_table.up.sql
@@ -9,6 +9,6 @@ CREATE TABLE orders (
     number VARCHAR PRIMARY KEY,
     user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     status order_status NOT NULL DEFAULT 'NEW',
-    accrual INTEGER,
+    accrual DECIMAL(10,2),
     uploaded_at TIMESTAMP NOT NULL DEFAULT NOW()
 );
